backend/AI_TMP/ai/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "consumption-poc"
version = "0.1.0"
description = "AI consumption report POC"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "google-genai>=1.0,<2.0",
  "elevenlabs>=1.8,<2.0",
  "python-dotenv>=1.0,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.2,<3.0",
  "scikit-learn>=1.5,<2.0",
  "lightgbm>=4.5,<5.0",
  "catboost>=1.2,<2.0",
  "dice-ml>=0.11,<1.0",
  "shap>=0.46,<  1.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.0,<9.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
